feat(dashboard): Improve DreTable month display
Dynamically adjusts the displayed months in the DRE table based on the selected period. This prevents displaying future months when a specific historical period is chosen, improving clarity and relevance of the data presented.

Also, refactors the EstiloLinha component to include new styling properties and options for finer control over text and background appearance across different views (screen and document).